+++ sandbox/index.js
@@ -5,10 +5,25 @@ window.MCDatepicker = MCDatepicker;
 
 const datepickerBTN = document.querySelector('#datepicker_btn');
 const inputFourDatepicker = document.querySelector('#datepicker_four');
+const setDateInput = document.querySelector('#set_date_input_two');
+const setDayInput = document.querySelector('#set_day_input_two');
+const setMonthInput = document.querySelector('#set_month_input_two');
+const setYearInput = document.querySelector('#set_year_input_two');
+const setDateBtn = document.querySelector('#set_date_btn');
+const setDayBtn = document.querySelector('#set_day_btn');
+const setMonthBtn = document.querySelector('#set_month_btn');
+const setYearBtn = document.querySelector('#set_year_btn');
+
+let setDateValue = null;
+let setDayValue = null;
+let setMonthValue = null;
+let setYearValue = null;
 
 const firstDatePicker = MCDatepicker.create({
 	el: '#datepicker_one',
 	bodyType: 'inline',
+	autoClose: true,
+	closeOnBlur: true,
 	dateFormat: 'dddd, dd mmmm yyyy',
 	disableDates: [new Date(2021, 1, 12), new Date(2021, 4, 15)],
 	customWeekDays: ['Duminică', 'Luni', 'Marți', 'Miercuri', 'Joi', 'Vineri', 'Sâmbătă'],
@@ -46,7 +61,10 @@ firstDatePicker.onCancel(() => console.log('Cancel button clicked!'));
 
 const secundDatePicker = MCDatepicker.create({
 	el: '#datepicker_two',
+	// dateFormat: 'yyyy',
+	// dateFormat: 'mm-yyyy',
 	bodyType: 'modal',
+	// closeOndblclick: false,
 	// showCalendarDisplay: false,
 	minDate: new Date(2020, 11, 5),
 	maxDate: new Date(2021, 0, 16)
@@ -54,6 +72,8 @@ const secundDatePicker = MCDatepicker.create({
 const thirdDatePicker = MCDatepicker.create({
 	el: '#datepicker_three',
 	dateFormat: 'mm-yyyy',
+	autoClose: true,
+	// dateFormat: 'dd-mm-yyyy',
 	customWeekDays: ['Søndag', 'Måneder', 'Tirsdag', 'Onsdag', 'Torsdag', 'Fredag', 'Lørdag'],
 	customMonths: [
 		'Januar',
@@ -80,6 +100,8 @@ const thirdDatePicker = MCDatepicker.create({
 const forthDatePicker = MCDatepicker.create({
 	dateFormat: 'yyyy',
 	selectedDate: new Date('2020-04-04'),
+	closeOndblclick: false,
+	closeOnBlur: true,
 	allowedMonths: [5, 7, 9],
 	allowedYears: [2016, 2018, 2020, 2022, 2024, 2026]
 });
@@ -93,3 +115,45 @@ datepickerBTN.onclick = () => {
 };
 
 forthDatePicker.onCancel(() => alert('The datepicker was closed using CANCEL button'));
+
+// ---------------------
+
+setDateInput.onchange = (e) => {
+	// converts string into code ex: new Date(2021, 0, 1)
+	setDateValue = eval(e.target.value);
+};
+
+setDateBtn.onclick = () => {
+	setDateValue !== null && secundDatePicker.setFullDate(setDateValue);
+};
+
+// ---------------------
+
+setDayInput.onchange = (e) => {
+	setDayValue = Number(e.target.value);
+};
+
+setDayBtn.onclick = () => {
+	setDayValue !== null && secundDatePicker.setDate(setDayValue);
+};
+
+// ---------------------
+
+setMonthInput.onchange = (e) => {
+	setMonthValue = Number(e.target.value);
+};
+
+setMonthBtn.onclick = () => {
+	console.log(setMonthValue);
+	setMonthValue !== null && secundDatePicker.setMonth(setMonthValue);
+};
+
+// ---------------------
+
+setYearInput.onchange = (e) => {
+	setYearValue = Number(e.target.value);
+};
+
+setYearBtn.onclick = () => {
+	setYearValue !== null && secundDatePicker.setYear(setYearValue);
+};