+++ src/js/emiters.js
@@ -8,7 +8,8 @@ import {
 	CHANGE_MONTH,
 	CHANGE_YEAR,
 	DATE_PICK,
-	PREVIEW_PICK
+	PREVIEW_PICK,
+	SET_DATE
 } from './events';
 
 export const dispatchCalendarShow = (elem, instance) => {
@@ -42,11 +43,12 @@ export const dispatchHeaderUpdate = (elem) => {
 	elem.dispatchEvent(new CustomEvent(HEADER_UPDATE, { bubbles: true }));
 };
 
-export const dispatchDatePick = (elem) => {
+export const dispatchDatePick = (elem, dblclick = false) => {
 	elem.dispatchEvent(
 		new CustomEvent(DATE_PICK, {
 			bubbles: true,
 			detail: {
+				dblclick,
 				date: new Date(elem.getAttribute('data-val-date'))
 			}
 		})
@@ -71,15 +73,23 @@ export const dispatchChangeYear = (elem, direction) => {
 	);
 };
 
-export const dispatchPreviewCellPick = (elem) => {
-	const target = elem.offsetParent.getAttribute('data-target');
+export const dispatchPreviewCellPick = (elem, dblclick = false) => {
 	elem.dispatchEvent(
 		new CustomEvent(PREVIEW_PICK, {
 			bubbles: true,
 			detail: {
-				target,
+				dblclick,
 				data: elem.children[0].innerHTML
 			}
+		})
+	);
+};
+
+export const dispatchSetDate = (elem, detail = { instance: null, date: null }) => {
+	elem.dispatchEvent(
+		new CustomEvent(SET_DATE, {
+			bubbles: true,
+			detail
 		})
 	);
 };