+++ src/js/mc-calendar.js
@@ -1,10 +1,8 @@
-import { noop } from './utils';
 import defaultOptions from './defaults';
-import { applyOnFocusListener, removeOnFocusListener } from './listeners';
 import createInstance from './instance';
-import { validateOptions } from './validators';
 import { writeTemplate } from './render';
-import { dispatchCalendarShow, dispatchCalendarHide } from './emiters';
+import { validateOptions } from './validators';
+import { applyOnFocusListener, removeOnFocusListener } from './listeners';
 
 import '../scss/main.scss';
 
@@ -12,25 +10,28 @@ const MCDatepicker = (() => {
 	let datepickers = [];
 	let calendarNodes = null;
 
-	let initCalendar = () => {
-		initCalendar = noop;
-		calendarNodes = writeTemplate(datepickers);
+	const initCalendar = () => {
+		if (calendarNodes) return;
+		calendarNodes = writeTemplate();
 	};
 
-	const open = (el) => {
-		const { calendar } = calendarNodes;
-		dispatchCalendarShow(calendar, el);
+	const open = (instance) => {
+		const activeInstance =
+			datepickers.find((datepicker) => JSON.stringify(datepicker) === JSON.stringify(instance)) ||
+			null;
+		if (!activeInstance && !calendarNodes) return;
+		calendarNodes.calendarStates.open(activeInstance);
 	};
 
 	const close = () => {
-		const { calendar } = calendarNodes;
-		dispatchCalendarHide(calendar);
+		if (!calendarNodes) return;
+		const { calendarStates } = calendarNodes;
+		calendarStates.close();
 	};
 
 	const create = (options = {}) => {
 		// initiate the calendar instance once
 		initCalendar();
-		const { calendar } = calendarNodes;
 		// validate options and merge them with de default Options
 		const instanceOptions = validateOptions(options, defaultOptions);
 		// create instance
@@ -38,19 +39,22 @@ const MCDatepicker = (() => {
 		// push fresh created instance to instances array
 		datepickers.push(instance);
 		// add event listener to the linked input
-		if (options.hasOwnProperty('el')) {
-			applyOnFocusListener(calendar, instance);
-		}
+		applyOnFocusListener(instance);
 
 		return instance;
 	};
 	const remove = (instance) => {
+		if (!datepickers.length) return;
 		// remove the onFocus listener
-		if (options.hasOwnProperty('el')) {
-			removeOnFocusListener(instance);
-		}
+		removeOnFocusListener(instance);
 		// Remove the instance from the datepickers array
 		datepickers.splice(datepickers.indexOf(instance), 1);
+
+		if (!datepickers.length) {
+			const { calendar } = calendarNodes;
+			calendar.parentNode.removeChild(calendar);
+			calendarNodes = null;
+		}
 	};
 	return { create, remove, open, close };
 })();