+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.5.0
+
+---
+
+- added `autoClose` option
+- added `closeOnBlur` option
+- added `closeOndblclick` option
+- added `setFullDate()` method
+- added `setDate()` method
+- added `setMonth()` method
+- added `setYear()` method
+- fixed styles for safari
+- fixed animations for safari
+
 ## 0.4.0
 
 ---